Revert "Feat/110/paginate settings (#114)"
This reverts commit 776eff7bf6b9a9cc2569bbeabbbbb9eb7ca4648e.
