+++ tetris-client/src/main/resources/view/custom-setting-view.fxml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.layout.BorderPane?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.layout.Region?>
+
+
+<BorderPane xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1"
+            stylesheets="@../css/application.css, @../css/custom-setting-view.css"
+            fx:controller="seoultech.se.client.controller.CustomSettingSceneController">
+    <top>
+      <Label styleClass="title-label" text="Custom" BorderPane.alignment="CENTER" />
+   </top>
+   <center>
+      <VBox fx:id="settingContainer" styleClass="center-box" alignment="TOP_CENTER">
+         <children>
+            <Button fx:id="testSettingButton" onAction="#handleTestSettingButton" styleClass="menu-button-middle" text="*setting name*" />
+         </children>
+      </VBox>
+   </center>
+   <bottom>
+      <HBox styleClass="bottom-box, bottom-box-custom">
+         <children>
+            <Button fx:id="saveButton" onAction="#handleSaveButton" styleClass="menu-button-small" text="Save" />
+            <Button fx:id="deleteButton" onAction="#handleDeleteButton" styleClass="menu-button-small" text="Delete" />
+            <Region HBox.hgrow="ALWAYS"/>
+            <Button fx:id="backButton" onAction="#handleBackButton" styleClass="menu-button-small" text="Back" />
+         </children>
+      </HBox>
+   </bottom>
+</BorderPane>
